pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "levy"
version = "0.1.0"
description = "Semantic caching engine for LLM APIs – experimental core for cost optimisation"
authors = [{name = "AlejoJamC", email = "alejojamc7@gmail.com"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.24.0",
    "numpy>=1.24.0",
    "sentence-transformers>=2.2.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "mypy"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
